+++ docker_sql/api/login.php
@@ -21,7 +21,7 @@ if($conn){
             }
             $sqlUpdate = "UPDATE users SET apiKey = '" . $apiKey . "' WHERE username = '" . $username . "'";     
             if(mysqli_query($conn, $sqlUpdate)){
-                 $result = array("status" => "success", "message" => "Login successful", "username" => $row["username"], "apiKey" => $row["apiKey"]);
+                 $result = array("status" => "success", "message" => "Login successful", "username" => $row["username"], "apiKey" => $row["apiKey"] , "job" => $row["job"]);
             } else $result = array("status" => "failed", "message" => "Login failed try again");
         } else $result = array("status" => "failed", "message" => "Retry with correct email and password");
     }
